requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.61.1
    # via matplotlib
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.9.0
    # via -r requirements.in
numpy==2.0.0
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   opensimplex
opensimplex==0.4.5.1
    # via -r requirements.in
packaging==25.0
    # via matplotlib
perlin-noise==1.12
    # via -r requirements.in
pillow==10.4.0
    # via
    #   -r requirements.in
    #   matplotlib
psutil==6.0.0
    # via -r requirements.in
pyparsing==3.2.5
    # via matplotlib
pyqt6==6.7.0
    # via -r requirements.in
pyqt6-qt6==6.7.3
    # via pyqt6
pyqt6-sip==13.10.3
    # via pyqt6
python-dateutil==2.9.0.post0
    # via matplotlib
six==1.17.0
    # via python-dateutil
